backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic-settings>=2.6.0
qdrant-client>=1.12.0
google-genai>=1.0.0
sentence-transformers>=3.0.0
